rename fragment_tree folder to fragment_chain
